mos-prototype_redis/requirements.txt
async-timeout==4.0.2
Deprecated==1.2.13
grpcio==1.44.0
grpcio-tools==1.44.0
packaging==21.3
protobuf==3.20.0
pyparsing==3.0.8
redis==4.2.2
six==1.16.0
toml==0.10.2
wrapt==1.14.0
